backend/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-ninja==1.4.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ffmpeg-python==0.2.0
filelock==3.18.0
fsspec==2025.7.0
future==1.0.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
opencv-python-headless==4.12.0.88
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yt-dlp==2025.6.30
